feat: provide `faucet add` command
A new `faucet` subcommand is introduced, enabling the faucet to be controlled as a service. This is
necessary because as with nodes, the faucet also needs to be upgraded on a new deployment; it may
also require downgrading in the event of a problem.

The existing `faucet` command, which was created as a utility for the backwards compatibility test,
is removed; now that we can define the faucet as a service, it will no longer be necessary.

The function that determines the data directory for the faucet was being duplicated in two different
crates, and the node manager would have added a third, so it was refactored as public in
`sn_transfers` and shared from there.